requirements.txt
discord-py-interactions
interactions-tasks
selenium
webdriver_manager
attrs
pymongo
motor